+++ action/chaos_test.go
@@ -470,12 +470,12 @@ func TestNetemDealyByName(t *testing.T) {
 		Duration:     1 * time.Second,
 		Time:         120,
 		Jitter:       25,
-		Correlation:  15,
+		Correlation:  0.23,
 	}
 	client := container.NewMockSamalbaClient()
 	client.On("ListContainers", mock.AnythingOfType("container.Filter")).Return(cs, nil)
 	for _, c := range cs {
-		client.On("NetemContainer", c, "eth1", "delay 120ms 25ms 15%", net.ParseIP(""), 1*time.Second).Return(nil)
+		client.On("NetemContainer", c, "eth1", "delay 120ms 25ms 0.23", net.ParseIP(""), 1*time.Second).Return(nil)
 		client.On("StopNetemContainer", c, "eth1").Return(nil)
 	}
 	// do action
@@ -495,11 +495,11 @@ func TestNetemDealyByNameRandom(t *testing.T) {
 		Duration:     1 * time.Second,
 		Time:         120,
 		Jitter:       25,
-		Correlation:  15,
+		Correlation:  5.5,
 	}
 	client := container.NewMockSamalbaClient()
 	client.On("ListContainers", mock.AnythingOfType("container.Filter")).Return(cs, nil)
-	client.On("NetemContainer", mock.AnythingOfType("container.Container"), "eth1", "delay 120ms 25ms 15%", net.ParseIP(""), 1*time.Second).Return(nil)
+	client.On("NetemContainer", mock.AnythingOfType("container.Container"), "eth1", "delay 120ms 25ms 5.50", net.ParseIP(""), 1*time.Second).Return(nil)
 	client.On("StopNetemContainer", mock.AnythingOfType("container.Container"), "eth1").Return(nil)
 	// do action
 	RandomMode = true
@@ -525,7 +525,7 @@ func TestNetemDealyByPattern(t *testing.T) {
 	client := container.NewMockSamalbaClient()
 	client.On("ListContainers", mock.AnythingOfType("container.Filter")).Return(cs, nil)
 	for _, c := range cs {
-		client.On("NetemContainer", c, "eth1", "delay 120ms 25ms 15%", net.ParseIP(""), 1*time.Second).Return(nil)
+		client.On("NetemContainer", c, "eth1", "delay 120ms 25ms 15.00", net.ParseIP(""), 1*time.Second).Return(nil)
 		client.On("StopNetemContainer", c, "eth1").Return(nil)
 	}
 	// do action
@@ -546,12 +546,12 @@ func TestNetemDealyByPatternIPFilter(t *testing.T) {
 		Duration:     1 * time.Second,
 		Time:         120,
 		Jitter:       25,
-		Correlation:  15,
+		Correlation:  10,
 	}
 	client := container.NewMockSamalbaClient()
 	client.On("ListContainers", mock.AnythingOfType("container.Filter")).Return(cs, nil)
 	for _, c := range cs {
-		client.On("NetemContainer", c, "eth1", "delay 120ms 25ms 15%", ip, 1*time.Second).Return(nil)
+		client.On("NetemContainer", c, "eth1", "delay 120ms 25ms 10.00", ip, 1*time.Second).Return(nil)
 		client.On("StopNetemContainer", c, "eth1").Return(nil)
 	}
 	// do action
@@ -571,11 +571,11 @@ func TestNetemDealyByPatternRandom(t *testing.T) {
 		Duration:     1 * time.Second,
 		Time:         120,
 		Jitter:       25,
-		Correlation:  15,
+		Correlation:  10.2,
 	}
 	client := container.NewMockSamalbaClient()
 	client.On("ListContainers", mock.AnythingOfType("container.Filter")).Return(cs, nil)
-	client.On("NetemContainer", mock.AnythingOfType("container.Container"), "eth1", "delay 120ms 25ms 15%", net.ParseIP(""), 1*time.Second).Return(nil)
+	client.On("NetemContainer", mock.AnythingOfType("container.Container"), "eth1", "delay 120ms 25ms 10.20", net.ParseIP(""), 1*time.Second).Return(nil)
 	client.On("StopNetemContainer", mock.AnythingOfType("container.Container"), "eth1").Return(nil)
 	// do action
 	RandomMode = true